Move sheet templates to src/gatsby-templates
