+++ src/web/components/CommentsArea.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
+import { DisqusData } from '../../types/types'
 import DisqusScript from '../providers/DisqusScript'
 import { RenderProps } from '../providers/DisqusScript'
-import { DisqusData } from '../types'
 import CSS from './CommentsArea.module.scss'
 import CommentsAreaSummary from './CommentsAreaSummary'
 import CommentsSection from './CommentsSection'
