+++ src/web/components/PagesList.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import useSiteContent from '../../gatsby-hooks/useSiteContent'
-import { SiteLink } from '../types'
+import { SiteLink } from '../../types/types'
 import PageLink from './PageLink'
 import CSS from './PagesList.module.scss'
