+++ src/web/components/PostContent/transform.tsx
@@ -7,7 +7,7 @@ import RehypeReact from 'rehype-react'
 import separatify from 'rehype-table-separators'
 // @ts-ignore TODO types for rehype-wrapify
 import wrapify from '../../../rehype-wrapify'
-import { HastNode } from '../../types'
+import { HastNode } from '../../../types/types'
 
 /**
  * Converts a HAST node (generated from Markdown) into a React element.
