+++ src/web/components/RelatedPostsArea.tsx
@@ -1,6 +1,6 @@
 /* @flow */
 import * as React from 'react'
-import { SiteLink } from '../types'
+import { SiteLink } from '../../types/types'
 import RelatedPostsSection from './RelatedPostsSection'
 
 /**
