+++ src/web/components/RootPage.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import Helmet from 'react-helmet'
 
 import useSiteContent from '../../gatsby-hooks/useSiteContent'
+import { GroupedSiteLinks, SiteLink } from '../../types/types'
 import { LiveSearchInput } from '../../web-search'
-import { GroupedSiteLinks, SiteLink } from '../types'
 import CommonHead from './CommonHead'
 import PagesList from './PagesList'
 import SiteHeader from './SiteHeader'
