+++ src/web/lib/site_page.tsx
@@ -1,5 +1,5 @@
 import groupBy from 'group-by'
-import { AllSitePage, GroupedSiteLinks, PageEdge, SiteLink } from '../types'
+import { AllSitePage, GroupedSiteLinks, PageEdge, SiteLink } from '../../types/types'
 
 /**
  * Converts AllSitePage to SiteLinks
